# Proof of Concept: VAES Support