The idea is to add use VAES instruction to scan wider length each loop iteration.
(I also tested scan the same length per loop with less instruction, but it does not speed up at all).

We can gain 100% speed up.

Without `VAES`
```
aeshash/string          time:   [59.425 ns 59.475 ns 59.523 ns]                           
aeshash/wider-string    time:   [1.2053 µs 1.2062 µs 1.2075 µs]   
```

With `VAES`:
```
aeshash/string          time:   [52.435 ns 52.511 ns 52.604 ns]                           
                        change: [-11.669% -11.523% -11.383%] (p = 0.00 < 0.05)
                        Performance has improved.
aeshash/wider-string    time:   [531.93 ns 532.18 ns 532.47 ns]                                  
                        change: [-55.856% -55.809% -55.765%] (p = 0.00 < 0.05)
                        Performance has improved.
```

-----------------------

**Notice:**
- `aeshash/wider-string` is same as `aeshash/string` except that its lengths set has larger data point.
- `vaes` passed all quality tests but its hash value may not be compatible with non-aes targets.