@SchrodingerZhu 2022-11-06
SMHasher FAILS