@SchrodingerZhu 2022-11-06
All SMHasher tests passed with latest fix