@SchrodingerZhu 2022-11-06
On EPYC 7773X, one can get the following results with this patch:
```
# AHASH64
Average      - 34.263 bytes/cycle - 98028.08 MiB/sec @ 3 ghz
# WYHASH
Average      - 10.938 bytes/cycle - 31293.76 MiB/sec @ 3 ghz
# XXH3
Average      - 20.607 bytes/cycle - 58957.09 MiB/sec @ 3 ghz
# T1HA0_AES_AVX2
Average      - 32.612 bytes/cycle - 93304.43 MiB/sec @ 3 ghz
```