@tkaitchuck 2022-11-10
Thanks for putting this together.
Would you like to collaborate working in a single branch?
I don't have such a CPU to test on, which has held me back on this feature.